decompile unk_020932A4
Functions:
PhoneBookTrainerGetRematchInfo

Had to modify TryGetRematchTrainerIdByBaseTrainerId's return type

+++ include/overlay_26.h
@@ -3,7 +3,7 @@
 
 #include "script.h"
 
-u32 TryGetRematchTrainerIdByBaseTrainerId(SAVEDATA* savedata, u16 trainer_no);
+u16 TryGetRematchTrainerIdByBaseTrainerId(SAVEDATA* savedata, u16 trainer_no);
 void ov26_022599D8(FieldSystem *fsys, LocalMapObject *lastTalked);
 
 #endif //POKEHEARTGOLD_OVERLAY_26_H
